refactor: migrate to "stack" package
